Add support for Anka Build 2.3 and Big Sur
If we cannot detect the FUSE driver within the VM, we will use the
`anka cp` utility to copy the contents of the volume (or current working
directory) into the VM at `/private/var/tmp/ankafs.0`. We do this to
try and maintain backwards compatibility for jobs that may expect
content to exist at that location.

In addition, we have split up the logic that was in just the `command`
hook into several other hooks throughout the job lifecycle. This makes
it easier to unit test, but also provides some minor benefits in terms
of optimizing behavior.

Signed-off-by: Tom Duffield <tom@chef.io>